Remove warnings

+++ src/io/model/modelreadermultiformat.cpp
@@ -44,8 +44,8 @@ namespace {
     {
         for (unsigned int i = 0; i < scene->mNumMeshes; ++i) {
             meshArray.push_back(scene->mMeshes[i]);
-            totalSizeIndex += meshArray.back()->mNumFaces * 3;
-            totalSizeVertex += meshArray.back()->mNumVertices;
+            totalSizeIndex += static_cast<size_t>(meshArray.back()->mNumFaces) * 3;
+            totalSizeVertex += static_cast<size_t>(meshArray.back()->mNumVertices);
         }
 
         /*for (auto m = 0; m < currNode->mNumMeshes; ++m)
